requirements.txt
# This file was autogenerated by uv via the following command:
#    uv export -o requirements.txt --no-dev --no-emit-project --no-hashes
annotated-types==0.7.0
    # via pydantic
anyio==4.11.0
    # via httpx
attrs==25.4.0
    # via
    #   frictionless
    #   jsonlines
    #   jsonschema
    #   referencing
beautifulsoup4==4.14.2
    # via wikipedia
certifi==2025.11.12
    # via
    #   httpcore
    #   httpx
    #   pyogrio
    #   pyproj
    #   requests
chardet==5.2.0
    # via frictionless
charset-normalizer==3.4.4
    # via requests
ckanapi==4.9
    # via hdx-python-api
click==8.3.0
    # via typer
colorama==0.4.6 ; sys_platform == 'win32'
    # via
    #   click
    #   defopt
    #   loguru
    #   tqdm
defopt==7.0.0
    # via hdx-python-api
dnspython==2.8.0
    # via email-validator
docopt==0.6.2
    # via
    #   ckanapi
    #   num2words
docutils==0.22.3
    # via defopt
email-validator==2.3.0
    # via hdx-python-api
et-xmlfile==2.0.0
    # via openpyxl
frictionless==5.18.1
    # via hdx-python-utilities
geopandas==1.1.1
    # via hdx-scraper-cod-ab-global
h11==0.16.0
    # via httpcore
h2==4.3.0
    # via httpx
hdx-python-api==6.5.2
    # via hdx-scraper-cod-ab-global
hdx-python-country==3.9.8
    # via
    #   hdx-python-api
    #   hdx-scraper-cod-ab-global
hdx-python-utilities==3.9.5
    # via
    #   hdx-python-api
    #   hdx-python-country
    #   hdx-scraper-cod-ab-global
hpack==4.1.0
    # via h2
httpcore==1.0.9
    # via httpx
httpx==0.28.1
    # via hdx-scraper-cod-ab-global
humanize==4.14.0
    # via frictionless
hyperframe==6.1.0
    # via h2
idna==3.11
    # via
    #   anyio
    #   email-validator
    #   httpx
    #   requests
ijson==3.4.0.post0
    # via hdx-python-utilities
inflect==7.5.0
    # via quantulum3
isodate==0.7.2
    # via frictionless
jinja2==3.1.6
    # via frictionless
joblib==1.5.2
    # via
    #   quantulum3
    #   scikit-learn
jsonlines==4.0.0
    # via hdx-python-utilities
jsonpath-ng==1.7.0
    # via libhxl
jsonschema==4.25.1
    # via
    #   frictionless
    #   tableschema-to-template
jsonschema-specifications==2025.9.1
    # via jsonschema
libhxl==5.2.2
    # via
    #   hdx-python-api
    #   hdx-python-country
loguru==0.7.3
    # via hdx-python-utilities
makefun==1.16.0
    # via hdx-python-api
markdown-it-py==4.0.0
    # via rich
marko==2.2.1
    # via frictionless
markupsafe==3.0.3
    # via jinja2
mdurl==0.1.2
    # via markdown-it-py
more-itertools==10.8.0
    # via inflect
num2words==0.5.14
    # via quantulum3
numpy==2.3.4
    # via
    #   geopandas
    #   pandas
    #   pyogrio
    #   quantulum3
    #   scikit-learn
    #   scipy
    #   shapely
openpyxl==3.1.5
    # via hdx-python-utilities
packaging==25.0
    # via
    #   geopandas
    #   pyogrio
pandas==2.3.3
    # via
    #   geopandas
    #   hdx-scraper-cod-ab-global
petl==1.7.17
    # via frictionless
ply==3.11
    # via
    #   jsonpath-ng
    #   libhxl
pockets==0.9.1
    # via sphinxcontrib-napoleon
pyarrow==22.0.0
    # via hdx-scraper-cod-ab-global
pydantic==2.12.4
    # via frictionless
pydantic-core==2.41.5
    # via pydantic
pygments==2.19.2
    # via rich
pyogrio==0.11.1
    # via geopandas
pyphonetics==0.5.3
    # via hdx-python-utilities
pyproj==3.7.2
    # via geopandas
python-dateutil==2.9.0.post0
    # via
    #   frictionless
    #   hdx-python-utilities
    #   libhxl
    #   pandas
python-dotenv==1.2.1
    # via hdx-scraper-cod-ab-global
python-io-wrapper==0.3.1
    # via libhxl
python-slugify==8.0.4
    # via
    #   ckanapi
    #   frictionless
pytz==2025.2
    # via pandas
pyyaml==6.0.3
    # via
    #   frictionless
    #   tableschema-to-template
quantulum3==0.9.2
    # via
    #   hdx-python-api
    #   hdx-scraper-cod-ab-global
ratelimit==2.2.1
    # via hdx-python-utilities
referencing==0.37.0
    # via
    #   jsonschema
    #   jsonschema-specifications
requests==2.32.5
    # via
    #   ckanapi
    #   frictionless
    #   hdx-python-api
    #   libhxl
    #   requests-file
    #   wikipedia
requests-file==3.0.1
    # via hdx-python-utilities
rfc3986==2.0.0
    # via frictionless
rich==14.2.0
    # via typer
rpds-py==0.28.0
    # via
    #   jsonschema
    #   referencing
ruamel-yaml==0.18.16
    # via hdx-python-utilities
ruamel-yaml-clib==0.2.14 ; platform_python_implementation == 'CPython'
    # via ruamel-yaml
scikit-learn==1.7.2
    # via quantulum3
scipy==1.16.3
    # via
    #   quantulum3
    #   scikit-learn
setuptools==80.9.0
    # via ckanapi
shapely==2.1.2
    # via
    #   geopandas
    #   hdx-scraper-cod-ab-global
shellingham==1.5.4
    # via typer
simpleeval==1.0.3
    # via frictionless
simplejson==3.20.2
    # via ckanapi
six==1.17.0
    # via
    #   ckanapi
    #   pockets
    #   python-dateutil
    #   sphinxcontrib-napoleon
sniffio==1.3.1
    # via anyio
soupsieve==2.8
    # via beautifulsoup4
sphinxcontrib-napoleon==0.7
    # via defopt
stemming==1.0.1
    # via quantulum3
structlog==25.5.0
    # via libhxl
tableschema-to-template==0.0.13
    # via hdx-python-utilities
tabulate==0.9.0
    # via frictionless
tenacity==9.1.2
    # via
    #   hdx-python-country
    #   hdx-scraper-cod-ab-global
text-unidecode==1.3
    # via python-slugify
threadpoolctl==3.6.0
    # via scikit-learn
tqdm==4.67.1
    # via hdx-scraper-cod-ab-global
typeguard==4.4.4
    # via inflect
typer==0.20.0
    # via frictionless
typing-extensions==4.15.0
    # via
    #   beautifulsoup4
    #   frictionless
    #   pydantic
    #   pydantic-core
    #   typeguard
    #   typer
    #   typing-inspection
typing-inspection==0.4.2
    # via pydantic
tzdata==2025.2
    # via pandas
unidecode==1.4.0
    # via
    #   libhxl
    #   pyphonetics
urllib3==2.5.0
    # via
    #   libhxl
    #   requests
validators==0.35.0
    # via frictionless
wheel==0.45.1
    # via libhxl
wikipedia==1.4.0
    # via quantulum3
win32-setctime==1.2.0 ; sys_platform == 'win32'
    # via loguru
xlrd==2.0.2
    # via hdx-python-utilities
xlrd3==1.1.0
    # via libhxl
xlsx2csv==0.8.4
    # via hdx-python-utilities
xlsxwriter==3.2.9
    # via tableschema-to-template
xlwt==1.3.0
    # via hdx-python-utilities
